Ensure correct shutdown of an aborted archive listing.
Because defers execute LIFO, the wait for status after the process starts will
happen before the previously-deferred cleanup cancellation. We still want that
one to ensure the context gets cleaned up on earlier exits; but we also need to
ensure the context is signalled BEFORE waiting on the happy path.